requirements.txt
Pillow==8.3.1
pyglet==1.5.18